podcast-generator/requirements.txt
# Podcast Generator Requirements

# Google AI SDK
google-generativeai>=0.3.0

# Audio processing
pydub>=0.25.1

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# For production Google Cloud TTS
google-cloud-texttospeech>=2.14.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0